rgw: move manifest code around, initial tier separation
cascading changes, and minor improvements.
