+++ src/rgw/librgw.cc
@@ -54,6 +54,8 @@
 #include "rgw_http_client.h"
 #include "rgw_http_client_curl.h"
 
+#include "services/svc_zone.h"
+
 #include <errno.h>
 #include <thread>
 #include <string>
