+++ src/rgw/rgw_cr_tools.cc
@@ -7,6 +7,8 @@
 #include "rgw_acl_s3.h"
 #include "rgw_zone.h"
 
+#include "services/svc_zone.h"
+
 #define dout_context g_ceph_context
 #define dout_subsys ceph_subsys_rgw
