+++ src/rgw/rgw_lc.h
@@ -18,7 +18,6 @@
 #include "common/Thread.h"
 #include "rgw_common.h"
 #include "rgw_rados.h"
-#include "rgw_multi.h"
 #include "cls/rgw/cls_rgw_types.h"
 #include "rgw_tag.h"
