+++ src/rgw/rgw_putobj_processor.h
@@ -20,6 +20,7 @@
 #include "rgw_putobj.h"
 #include "rgw_rados.h"
 #include "services/svc_rados.h"
+#include "services/svc_tier_rados.h"
 
 namespace rgw {
