+++ src/rgw/rgw_tools.cc
@@ -21,6 +21,7 @@
 #include "rgw_zone.h"
 
 #include "services/svc_sys_obj.h"
+#include "services/svc_zone.h"
 #include "services/svc_zone_utils.h"
 
 #define dout_subsys ceph_subsys_rgw
